+++ hosts/echo/darwin-configuration.nix
@@ -27,4 +27,6 @@
     "/System/Volumes/Preboot/Cryptexes/App/System/Applications/Safari.app"
     "/System/Applications/System Settings.app"
   ];
+
+  system.stateVersion = 5;
 }