+++ modules/laptop.nix
@@ -62,12 +62,7 @@
       builtins.attrValues { inherit (pkgs) aldente; };
 
     launchd.user.agents.alDente = {
-      serviceConfig.ProgramArguments = [
-        "/bin/sh"
-        "-c"
-        ''
-          /bin/wait4path /nix/store &amp;&amp; exec "/Applications/Nix Apps/AlDente.app/Contents/MacOS/AlDente"''
-      ];
+      command = ''"/Applications/Nix Apps/AlDente.app/Contents/MacOS/AlDente"'';
       serviceConfig.RunAtLoad = true;
     };